requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.2
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
distro==1.8.0
entrypoints==0.4
filelock==3.13.1
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.19.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.154
llama-index==0.6.11
markdown-it-py==3.0.0
MarkupSafe @ file:///D:/bld/markupsafe_1695367436673/work
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyarrow==14.0.1
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
pypdf==3.9.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.29.0
rich==13.6.0
rpds-py==0.12.0
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.15
streamlit==1.19.0
streamlit-chat==0.0.2.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.29.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1692105915497/work
yarl==1.9.2
zipp==3.17.0
